refactor(typescript-client): namespace model helpers under models
