+++ docs/phoenix/sdk-api-reference/typescript/packages/phoenix-client/overview.mdx
@@ -39,7 +39,7 @@ That gives the agent version-matched docs plus the exact implementation and gene
 | `@arizeai/phoenix-client` | `createClient`, generated OpenAPI types, config helpers |
 | `@arizeai/phoenix-client/projects` | Project listing and retention-policy assignment |
 | `@arizeai/phoenix-client/prompts` | Prompt CRUD plus `toSDK` conversion |
-| `@arizeai/phoenix-client/modelProviders` | Server-backed built-in model-provider discovery |
+| `@arizeai/phoenix-client/models` | Server-backed built-in model-provider discovery |
 | `@arizeai/phoenix-client/datasets` | Dataset creation and retrieval |
 | `@arizeai/phoenix-client/experiments` | Experiment execution and lifecycle |
 | `@arizeai/phoenix-client/spans` | Span search, notes, and span/document annotations |
@@ -176,7 +176,7 @@ Prefer this layer when:
     <li><code>src/types/core.ts</code></li>
     <li><code>src/projects/</code></li>
     <li><code>src/prompts/</code></li>
-    <li><code>src/modelProviders/</code></li>
+    <li><code>src/models/</code></li>
     <li><code>src/datasets/</code></li>
     <li><code>src/experiments/</code></li>
     <li><code>src/spans/</code></li>
