+++ js/packages/phoenix-client/test/entryPoints.test.ts
@@ -12,7 +12,7 @@ const cjsEntries = {
   jest: "../dist/src/jest/index.js",
   projects: "../dist/src/projects/index.js",
   users: "../dist/src/users/index.js",
-  modelProviders: "../dist/src/modelProviders/index.js",
+  models: "../dist/src/models/index.js",
 } as const;
 // The index entry transitively require()s @arizeai/phoenix-otel's dist via the
 // workspace symlink, so an unbuilt sibling would fail the guard even though
@@ -55,7 +55,7 @@ const esmEntries = {
   jest: "../dist/esm/jest/index.js",
   projects: "../dist/esm/projects/index.js",
   users: "../dist/esm/users/index.js",
-  modelProviders: "../dist/esm/modelProviders/index.js",
+  models: "../dist/esm/models/index.js",
 } as const;
 
 describe.skipIf(!isBuilt)("built ESM entry points", () => {
